+++ PJ3/Library/Bee/1900b0aE.dag.json
@@ -20564,7 +20564,7 @@
       "DisplayName": "Writing Assembly-CSharp.rsp",
       "ActionType": "WriteFile",
       "PayloadOffset": 1555269,
-      "PayloadLength": 33406,
+      "PayloadLength": 33446,
       "PayloadDebugContentSnippet": "-target:library\r\n-out:\"Library",
       "Inputs": [],
       "Outputs": [
@@ -20576,7 +20576,7 @@
       "Annotation": "WriteText Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.rsp2",
       "DisplayName": "Writing Assembly-CSharp.rsp2",
       "ActionType": "WriteFile",
-      "PayloadOffset": 1588762,
+      "PayloadOffset": 1588802,
       "PayloadLength": 0,
       "PayloadDebugContentSnippet": "",
       "Inputs": [],
@@ -20828,6 +20828,7 @@
         "Assets/Scripts/Managers/InspectionManager.cs",
         "Assets/Scripts/Managers/InteractionsManager.cs",
         "Assets/Scripts/Managers/InventoryManager.cs",
+        "Assets/Scripts/Managers/UIManager.cs",
         "Assets/Scripts/Objects/Gun1.cs",
         "Assets/Scripts/Player/MoveCamera.cs",
         "Assets/Scripts/Player/PlayerCam.cs",
@@ -21090,7 +21091,7 @@
       "Annotation": "WriteText Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.dll.mvfrm.rsp",
       "DisplayName": "Writing Assembly-CSharp.dll.mvfrm.rsp",
       "ActionType": "WriteFile",
-      "PayloadOffset": 1588858,
+      "PayloadOffset": 1588898,
       "PayloadLength": 9545,
       "PayloadDebugContentSnippet": "Library\\Bee\\artifacts\\mvdfrm\\U",
       "Inputs": [],
